requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
phonenumbers>=8.12.0
python-whois>=0.7.3
dnspython>=2.3.0
ipaddress>=1.0.23
Pillow>=9.0.0
hashlib>=1.0.0
Pillow>=9.0.0
python-dateutil>=2.8.0
urllib3>=1.26.0
certifi>=2022.0.0
